requirements.txt
# AI Universal Service Orchestrator - Core Dependencies
# Python 3.11+

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data validation & settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0
asyncpg>=0.29.0

# Resilience
tenacity>=8.2.0
circuitbreaker>=2.0.0

# Observability
opencensus-ext-azure>=1.1.0
opencensus-ext-logging>=0.1.0

# HTTP client
httpx>=0.26.0

# AI / LLM (Azure OpenAI - for Intent Resolver, Agentic AI)
openai>=1.12.0

# Optional: Google AI / Gemini (for Agentic AI planner when Azure not configured)
# google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
jinja2>=3.1.0

# Payment (Module 15)
stripe>=8.0.0

# Testing (server-based)
pytest>=7.4.0
pytest-asyncio>=0.23.0
